# .md/API-CONTRACT.yaml
openapi: 3.0.3
info:
  title: Turma do Rola - Comary — Contrato de API
  version: 0.12.0
  description: >-
    Contrato publicado incrementalmente, endpoint a endpoint, assim que cada
    contrato fica estável (TASK.md Seção 1.7 / GUARDRAILS.md regra implícita
    de rastreabilidade de API) — não é preciso esperar a tarefa de Backend
    inteira terminar. Frontend/Mobile começam a consumir assim que um
    endpoint aparece aqui, mesmo que a implementação completa da tarefa
    ainda esteja em andamento. Mudar um contrato já publicado é permitido,
    mas nunca silenciosamente — toda mudança de contrato incompatível
    incrementa `info.version` e é registrada na seção "Changelog" no final
    deste arquivo.

    Duas famílias de endpoint neste sistema:
      1. **PostgREST autogerado pelo Supabase** sobre as views públicas
         curadas (`ranking_publico`, `presenca_mensal_publica`) — consumido
         **diretamente** pelo Frontend com a chave `anon` do Supabase, sem
         passar pela camada de API própria (SDD.md Seção 7.5, ADR-005). É o
         único caso em que o Frontend fala com o Supabase sem intermediação
         de um Route Handler — decisão já tomada no SDD.md, não uma exceção
         nova. Documentado aqui (paths desta seção) porque também é
         superfície pública real, exposta à internet.
      2. **Route Handlers do Next.js (App Router)**, publicados a partir de
         BE-04 em diante — toda escrita e toda leitura autenticada da área
...
servers:
...
components:
  securitySchemes:
    supabaseAnonKey:
      type: apiKey
      in: header
      name: apikey
      description: >-
        Chave pública (`anon`) do Supabase — único credential usado pelas
        páginas públicas (T02/T03). Nunca a `service role` (TASK.md Seção
        1.2/1.7, GUARDRAILS.md regra 7).
    sessionCookie:
      type: apiKey
      in: cookie
      name: sessao_interna
      description: >-
        Cookie de sessão emitido por `POST /api/auth/login` (`httpOnly`,
        `Secure`, `SameSite=Strict`, TTL 8-12h — ADR-004, BE-04). Único
        mecanismo de autenticação da área interna (modelo binário, sem
        RBAC — SDD.md Seção 7.2): sessão válida = acesso total. Toda rota
        de escrita futura desta área (BE-06 em diante) deve declarar este
        esquema de segurança.
  responses:
    SessaoInvalida:
      description: >-
        Sem sessão válida (`middleware.ts`, BE-04, GUARDRAILS.md regra 17) —
        toda rota de escrita da área interna, exceto `/api/auth/login` e
        `/api/auth/logout`, retorna isto quando o cookie `sessao_interna`
        está ausente, expirado ou com assinatura inválida. Resposta
        reutilizável — toda rota de escrita futura (BE-06 em diante) deve
        referenciar `#/components/responses/SessaoInvalida` para o caso
        401, em vez de redefinir o formato.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErroSessaoInvalida"
  schemas:
...
    AtletaBody:
      type: object
      description: >-
        Corpo de `POST /api/atletas` e `PUT /api/atletas/{id}` (BE-06).
        Campos em `snake_case`, mesma convenção já usada pelas respostas de
        `RankingPublicoItem`/`PresencaMensalPublicaItem` no restante deste
        contrato (decisão de consistência, TASK.md Seção 1.0 — documentada
        na nota de status de BE-06 do `TASK.md`, não escalada).
      properties:
        nome_completo:
          type: string
          minLength: 1
        apelido_exibicao:
          type: string
          minLength: 1
          description: >-
            Opcional — quando ausente/em branco, o Backend deriva o primeiro
            "token" de `nome_completo` (RF-01.2/RN-06).
        contato:
          type: string
          minLength: 1
          nullable: true
        data_nascimento:
          type: string
          format: date
          description: >-
            Obrigatório neste formulário (RF-01.1, UX-SPEC.md T04 marca com
            `*`) — não pode ser no futuro. Coluna correspondente em
            `app.atleta` é `nullable` no banco só para acomodar registro
            migrado do legado sem essa informação (RF-08.3/BE-15), não para
            tornar o campo opcional aqui.
        consentimento_responsavel_obtido:
          type: boolean
          default: false
          description: >-
            Obrigatoriamente `true` quando `data_nascimento` resultar em
            idade < 18 anos na data da requisição — caso contrário a API
            recusa com `400` (RF-01.3/RN-02).
        pontuacao_inicial:
          type: number
          minimum: 0
        confirmar_duplicidade:
          type: boolean
          default: false
          description: >-
            Quando `nome_completo` coincide (normalizado — trim/espaço/caixa)
            com o de outro atleta ativo e este campo não vier `true`, a API
            recusa com `409` (alerta de duplicidade, RF-01.5) em vez de
            gravar. O Frontend reenvia a mesma requisição com
            `confirmar_duplicidade: true` depois que o organizador confirmar
            no modal (UX-SPEC.md T04) — nunca um bloqueio definitivo, não há
            `UNIQUE` em `nome_completo` no schema (BE-02).
      required: [nome_completo, data_nascimento, pontuacao_inicial]
      additionalProperties: false
    AtletaResponse:
      type: object
      description: >-
        Resposta de `GET/POST /api/atletas` (lista/criação) e
        `GET/PUT /api/atletas/{id}` (BE-06). Nunca devolvida a cliente sem
        sessão válida (RN-01/GUARDRAILS.md regra 19, ver tag `interno`).
      properties:
        id:
          type: string
          format: uuid
        nome_completo:
          type: string
        apelido_exibicao:
          type: string
        contato:
          type: string
          nullable: true
        data_nascimento:
          type: string
          format: date
          nullable: true
        consentimento_responsavel_obtido:
          type: boolean
        pontuacao_inicial:
          type: number
        ativo:
          type: boolean
        anonimizado_em:
          type: string
          format: date-time
          nullable: true
          description: >-
            Preenchido pela função `anonimizar_atleta` (ADR-011, BE-07) —
            ver `POST /api/atletas/{id}/anonimizar`.
        criado_em:
          type: string
          format: date-time
        nivel_tecnico:
          type: number
          description: >-
            RN-03 — média de pontos de eventos de jogo (gol/cartão) por
            rodada com presença (`app.atleta_nivel_tecnico`, migration
            BE-06). Fallback = `pontuacao_inicial` para atleta sem nenhuma
            presença registrada. Nunca um campo de entrada do formulário
            (RF-01.4 — sempre derivado).
        rodadas_presentes:
          type: integer
          description: Denominador usado no cálculo de `nivel_tecnico` acima.
      required:
        [
          id,
          nome_completo,
          apelido_exibicao,
          contato,
          data_nascimento,
          consentimento_responsavel_obtido,
          pontuacao_inicial,
          ativo,
          anonimizado_em,
          criado_em,
          nivel_tecnico,
          rodadas_presentes,
        ]
      additionalProperties: false
    ErroValidacaoAtleta:
      type: object
      description: >-
        Corpo de erro `400` de `POST/PUT /api/atletas*` — corpo malformado ou
        que falha `AtletaBody` (inclui o bloqueio de RF-01.3: menor de 18 sem
        `consentimento_responsavel_obtido`, reportado como um item de
        `detalhes` com `path: ["consentimento_responsavel_obtido"]`).
      properties:
        error:
          type: string
          example: "Requisição inválida."
        detalhes:
          type: array
          items:
            type: object
            properties:
              path:
                type: array
                items: {}
              message:
                type: string
            required: [path, message]
      required: [error]
      additionalProperties: false
    ErroDuplicidadeAtleta:
      type: object
      description: >-
        Corpo de erro `409` de `POST/PUT /api/atletas*` (RF-01.5) — emitido
        quando `nome_completo` (normalizado) coincide com o de outro atleta
        ativo e a requisição não trouxe `confirmar_duplicidade: true`.
      properties:
        error:
          type: string
          example: "duplicidade"
        atletas_duplicados:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              nome_completo:
                type: string
            required: [id, nome_completo]
      required: [error, atletas_duplicados]
      additionalProperties: false
...
    RodadaResponse:
      type: object
      description: Resposta de sucesso (`201`) de `POST /api/rodadas` (BE-08).
      properties:
        id:
          type: string
          format: uuid
        data:
          type: string
          format: date
        status:
          type: string
          enum: [lancada, excluida]
          description: Sempre `"lancada"` na resposta deste endpoint (soft-delete, BE-02/BE-09).
        criado_em:
          type: string
          format: date-time
        participacoes:
          type: array
          items:
            $ref: "#/components/schemas/ParticipacaoRodadaResponse"
      required: [id, data, status, criado_em, participacoes]
      additionalProperties: false
...
    RodadaResumoItem:
      type: object
      description: >-
        Um item de `GET /api/rodadas` (BE-16, T06 do `UX-SPEC.md` —
        "Histórico de Rodadas"). Lacuna deixada por BE-08/BE-09/BE-10
        (nenhuma tarefa anterior cobria LEITURA de listagem de rodada — BE-08
        documentou a omissão explicitamente na época; BE-09/BE-10 cobriram
        só log de auditoria e preview de correção, nunca a listagem em si) —
        fechada nesta versão por decisão explícita do usuário ao identificar
        o bloqueio de `FE-06`. `presentes` é o único campo além dos já
        publicados por `RodadaResponse`/`RodadaExcluidaResponse`
        (`id`/`data`/`status`/`criado_em`), literal do wireframe T06
        ("19/09/2026 · 18 presentes").
      properties:
        id:
          type: string
          format: uuid
        data:
          type: string
          format: date
        status:
          type: string
          enum: [lancada, excluida]
          description: >-
            Rodada `excluida` (soft-delete) aparece normalmente na lista,
            com o status visível — NUNCA escondida silenciosamente (decisão
            de detalhe, mesmo racional já usado para `restricao_obrigatoria`
            desativada em `RestricaoObrigatoriaResponse`/T10, BE-12: o
            histórico permanece visível, quem decide o que fazer com uma
            rodada excluída na tela é o Frontend, não uma omissão do
            Backend).
        criado_em:
          type: string
          format: date-time
        presentes:
          type: integer
          description: >-
            Contagem de `participacao_rodada.status = "presente"` desta
            rodada (nunca inclui `lesionado`/`ausente` — mesma restrição já
            usada por `RankingPublicoItem.presencas`, BE-03/BLOCKER-005,
            embora aqui sem o filtro de `status = "lancada"` da rodada, já
            que rodada `excluida` também aparece nesta listagem).
      required: [id, data, status, criado_em, presentes]
      additionalProperties: false
    ErroConsultaRodadas:
      type: object
      description: >-
        Corpo de erro `400` de `GET /api/rodadas` — `limit` inválido (não
        numérico, ≤ 0, ou acima do teto máximo de 200), mesmo formato de
        `ErroConsultaLogAuditoria` (BE-09).
      properties:
        error:
          type: string
          example: "Requisição inválida."
        detalhes:
          type: array
          items:
            type: object
            properties:
              path:
                type: array
                items: {}
              message:
                type: string
            required: [path, message]
      required: [error]
      additionalProperties: false
...
paths:
...
  /api/atletas:
    get:
      tags: [interno]
      summary: Lista atletas (BE-06, RF-01)
      description: >-
        Exige sessão válida mesmo sendo leitura (ver tag `interno`) —
        devolve `contato`/`data_nascimento`. Cada item inclui o nível
        técnico (RN-03) já calculado (`nivel_tecnico`/`rodadas_presentes`).
        Sem paginação nesta versão (porte do grupo — dezenas de atletas, não
        milhares, TASK.md Seção 1.0).
      operationId: getAtletas
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Lista de atletas, ordenada por `nome_completo` asc.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AtletaResponse"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
    post:
      tags: [interno]
      summary: Cadastra um atleta (BE-06, RF-01.1)
      description: >-
        Inicializa o saldo acumulado do atleta com `pontuacao_inicial`
        (RF-01.1/RN-10). Bloqueia com `400` se `data_nascimento` resultar em
        idade < 18 anos e `consentimento_responsavel_obtido` não vier `true`
        (RF-01.3/RN-02). Alerta de duplicidade de `nome_completo` (RF-01.5)
        — ver `409` abaixo.
      operationId: postAtletas
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AtletaBody"
      responses:
        "201":
          description: Atleta criado. `nivel_tecnico` = `pontuacao_inicial` (nenhuma presença ainda, fallback RN-03).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AtletaResponse"
        "400":
          description: >-
            Corpo malformado/inválido — inclui o bloqueio de RF-01.3 (menor
            de 18 sem consentimento marcado) como um item de `detalhes` com
            `path: ["consentimento_responsavel_obtido"]`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroValidacaoAtleta"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
        "409":
          description: >-
            `nome_completo` (normalizado) coincide com o de outro atleta
            ativo e a requisição não trouxe `confirmar_duplicidade: true`
            (RF-01.5) — nenhum registro foi criado. Reenviar a mesma
            requisição com `confirmar_duplicidade: true` para confirmar e
            criar mesmo assim.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroDuplicidadeAtleta"
...
  /api/rodadas:
    get:
      tags: [interno]
      summary: Lista rodadas em ordem cronológica decrescente (BE-16, T06 do UX-SPEC.md)
      description: >-
        Lacuna deixada por BE-08/BE-09/BE-10 (nenhuma delas cobria LEITURA de
        listagem de rodada — BE-08 documentou a omissão explicitamente na
        época) — fechada nesta versão por decisão explícita do usuário ao
        identificar o bloqueio de `FE-06`. Lista `app.rodada` ordenada por
        `data desc` e, em caso de empate (duas rodadas na mesma data civil,
        cenário válido via `confirmar_duplicidade`, RF-02.8), por
        `criado_em desc` como desempate. NUNCA filtra por `status` — rodada
        `excluida` (soft-delete) aparece normalmente, com o status visível
        (ver `RodadaResumoItem`). Leitura interna protegida por sessão mesmo
        sendo `GET` (mesmo racional já usado por `GET /api/log-auditoria`,
        BE-09) — `middleware.ts`, `INTERNAL_READ_PROTECTED_PREFIXES`
        (`/api/rodadas` já estava na lista desde BE-13, nenhuma mudança
        necessária nesta tarefa).
      operationId: getRodadas
      security:
        - sessionCookie: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          description: >-
            Teto de linhas retornadas (decisão de detalhe, TASK.md Seção
            1.0 — não exigido literalmente pelo critério de aceite; mesmo
            racional/mesmos valores de `GET /api/log-auditoria`, BE-09:
            `app.rodada` cresce indefinidamente ao longo das temporadas, sem
            expurgo automático).
      responses:
        "200":
          description: Lista ordenada do mais recente ao mais antigo.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RodadaResumoItem"
        "400":
          description: "`limit` inválido (não numérico, ≤ 0, ou acima de 200)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroConsultaRodadas"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
    post:
      tags: [interno]
      summary: Lança uma rodada — presença, eventos de jogo e cálculo automático de pontos (BE-08, RF-02)
      description: >-
        Aciona a função PL/pgSQL `app.lancar_rodada`, que roda dentro de uma
        única transação (TASK.md Seção 1.2): grava `app.rodada`, uma
        `app.participacao_rodada` por atleta, `app.evento_jogo` por evento
        de gol/cartão e exatamente um `app.lancamento_pontos` por atleta
        (ledger append-only, ADR-006) com o saldo já calculado. Uma
        exceção em qualquer ponto do processamento reverte 100% do que já
        tinha sido gravado nesta chamada (RNF-10 — nunca estado parcial
        visível). Valor de pontos de cada evento sempre lido do vigente em
        `app.configuracao_pontuacao` na data da rodada (RN-05: presença
        `+2`, ausência `0`, lesionado pontua como presença, gol `+3`,
        cartão amarelo `-1`, cartão vermelho `-3`) — nunca hardcoded.
        Bloqueia com `400` se qualquer participação com `status: "ausente"`
        trouxer `eventos` não vazio (RF-02.6) — validado na própria camada
        de API (`zod`) e repetido como defesa em profundidade estrutural
        dentro da função PL/pgSQL. Alerta de rodada duplicada (RF-02.8) —
        ver `409` abaixo; rodada com `status: "excluida"` (soft-delete,
        BE-02) na mesma `data` nunca dispara esse alerta. Substituições
        (RF-06) não fazem parte deste endpoint — registradas depois, a
        partir de times já definidos (BE-13, fora do escopo de BE-08).
      operationId: postRodadas
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LancarRodadaBody"
      responses:
        "201":
          description: >-
            Rodada lançada. Cada item de `participacoes` inclui
            `pontos_delta` — o valor já efetivamente gravado em
            `app.lancamento_pontos` para aquele atleta nesta rodada (soma de
            presença/ausência + todos os eventos de jogo do atleta).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RodadaResponse"
        "400":
          description: >-
            Corpo malformado/inválido — inclui o bloqueio de RF-02.6 (evento
            de jogo para atleta ausente) como um item de `detalhes` com
            `path: ["participacoes", <índice>, "eventos"]`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroValidacaoRodada"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
        "409":
          description: >-
            Já existe uma rodada `lancada` (nunca `excluida`) com a mesma
            `data` e a requisição não trouxe `confirmar_duplicidade: true`
            (RF-02.8) — nenhuma rodada foi criada. Reenviar a mesma
            requisição com `confirmar_duplicidade: true` para confirmar e
            lançar mesmo assim.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroDuplicidadeRodada"
...
  /api/times/sugestao:
    post:
      tags: [interno]
      summary: >-
        Gera uma sugestão de divisão de times para os presentes informados
        (BE-11, RF-05.1 a RF-05.3, ADR-007/ADR-010)
      description: >-
        Heurística determinística de duas fases (ADR-007): (1) backtracking
        com poda, por componente conexo do grafo de restrições obrigatórias
        ativas (union-find, ADR-010), garantindo 100% das restrições
        obrigatórias entre os presentes OU provando que nenhuma divisão em
        `N` times as satisfaz simultaneamente (RF-05.1/RF-05.2); (2) busca
        local (swap iterativo entre times) minimizando a diferença agregada
        de nível técnico (RN-03) + idade entre os times (RF-05.3,
        "melhor esforço", não necessariamente ótimo global). Não persiste
        nada em `app.time`/`app.time_atleta` — devolve só a SUGESTÃO
        (RF-05.4, "o organizador pode ajustar manualmente antes de
        confirmar"); persistir a divisão confirmada é um gap desta
        decomposição, sinalizado na nota de status de `BE-11` no `TASK.md`,
        não coberto por este endpoint. Guard de timeout de 8s (TASK.md
        Seção 6.2 item 3) — ver `500` abaixo.
      operationId: postSugestaoTimes
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SugestaoTimesBody"
      responses:
        "200":
          description: >-
            Geração concluída dentro do orçamento de tempo — `status: "ok"`
            (divisão válida, ver `SugestaoTimesOkResponse`) ou `status:
            "conflito"` (nenhuma divisão em `N` times satisfaz 100% das
            restrições obrigatórias, ver `SugestaoTimesConflitoResponse`).
            Os dois são resultados VÁLIDOS do algoritmo (RF-05.2) — nunca um
            erro HTTP.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/SugestaoTimesOkResponse"
                  - $ref: "#/components/schemas/SugestaoTimesConflitoResponse"
                discriminator:
                  propertyName: status
                  mapping:
                    ok: "#/components/schemas/SugestaoTimesOkResponse"
                    conflito: "#/components/schemas/SugestaoTimesConflitoResponse"
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroValidacaoSugestaoTimes"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
        "404":
          description: >-
            Algum item de `atletas_ids` não corresponde a nenhum atleta
            existente — mesmo contrato de erro de `POST/PUT /api/restricoes*`
            (BE-12), reaproveitado aqui por ser a mesma situação (id
            referenciado em corpo de requisição que não existe em
            `app.atleta`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroAtletaReferenciadoNaoEncontrado"
        "500":
          description: >-
            Execução acima do timeout configurado (TASK.md Seção 6.2 item 3)
            — falha técnica real, nunca a função serverless travada sem
            resposta.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroFalhaTecnicaMontagemTimes"
  /api/rodadas/{id}/times:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: id da rodada (`app.rodada`) a que a divisão de times pertence.
    post:
      tags: [interno]
      summary: >-
        Confirma/persiste a divisão de times de uma rodada (RF-05.4) —
        escopo AMPLIADO desta execução de BE-13, decisão explícita do
        usuário
      description: >-
        Resolve o GAP estrutural sinalizado na nota de status de `BE-11`:
        `POST /api/times/sugestao` nunca persiste `app.time`/
        `app.time_atleta` (só devolve a SUGESTÃO, RF-05.4). Este endpoint
        recebe a divisão já ajustada manualmente pelo organizador (formato
        de entrada equivalente ao `status: "ok"` de
        `POST /api/times/sugestao`, possivelmente reordenado/editado) e
        grava definitivamente em `app.time`/`app.time_atleta`, TODOS os
        times de uma rodada gravados juntos ou nenhum é (aciona
        `app.confirmar_times_rodada` via RPC, uma única transação Postgres
        — ver a migration
        `20260903160000_create_confirmar_times_rodada_function.sql` para a
        justificativa completa de usar uma função PL/pgSQL aqui, mesmo esta
        operação não alterar saldo do atleta: é a única forma de garantir
        atomicidade multi-tabela nesta stack baseada em PostgREST).
        Reconfirmar (chamar de novo para a mesma rodada) SUBSTITUI a
        divisão anterior por completo — salvo se já existir
        `app.substituicao` registrada contra os times atuais, caso em que a
        reconfirmação é bloqueada (`409`) para preservar a fidelidade
        histórica das substituições já registradas (RF-06.1). Pré-requisito
        para `POST /api/rodadas/{id}/substituicoes` (BE-13) ter um
        `time_id` válido para vincular.
      operationId: postConfirmarTimesRodada
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmarTimesBody"
      responses:
        "200":
          description: >-
            Divisão gravada (ou reconfirmada — substitui a anterior) com
            sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimesConfirmadosResponse"
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroValidacaoConfirmarTimes"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
        "404":
          description: >-
            Rodada (`id` da URL) não encontrada (`ErroRodadaNaoEncontrada`,
            já publicado por BE-09) ou algum item de `times[].atletas_ids`
            não corresponde a nenhum atleta existente
            (`ErroAtletaReferenciadoNaoEncontrado`, já publicado por
            BE-12).
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErroRodadaNaoEncontrada"
                  - $ref: "#/components/schemas/ErroAtletaReferenciadoNaoEncontrado"
        "409":
          description: >-
            Rodada já excluída (`ErroRodadaJaExcluida`, já publicado por
            BE-09, mesma condição de fundo `errcode RD001`) ou já existe
            substituição registrada contra a divisão atual
            (`ErroSubstituicaoExistenteBloqueiaReconfirmacao`, `errcode
            TM001`, novo nesta versão).
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErroRodadaJaExcluida"
                  - $ref: "#/components/schemas/ErroSubstituicaoExistenteBloqueiaReconfirmacao"
...